requirements.txt
streamlit>=1.38.0
pandas>=2.2.0
plotly>=5.22.0
openpyxl>=3.1.2
python-dateutil>=2.9.0
openai>=1.0.0  # For OpenAI API
python-dotenv>=1.0.0  # For loading .env file
streamlit-openai>=0.1.0  # For chat interface with OpenAI

